Move method.ecma376_encrypted to method.container.ecma376_encrypted
