+++ msoffcrypto/method/container/ecma376_encrypted.py
